requirements.txt
streamlit
pydub
speechrecognition
streamlit-mic-recorder
google-generativeai
google-genai
google-cloud-texttospeech
streamlit-mic-recorder
google-ai-generativelanguage
numpy
pandas
scipy
requests
sounddevice
httpx
pydantic
typing-inspection
